requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
plotly==5.19.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1 